+++ src/app/dataset-view/additional-components/data-component/data.component.spec.ts
@@ -28,6 +28,7 @@ import { MatProgressBarModule } from "@angular/material/progress-bar";
 import { OverviewUpdate } from "../../dataset.subscriptions.interface";
 import _ from "lodash";
 import { RequestTimerComponent } from "./request-timer/request-timer.component";
+import { SqlEditorComponent } from "src/app/shared/editor/components/sql-editor/sql-editor.component";
 
 describe("DataComponent", () => {
     let component: DataComponent;
@@ -46,8 +47,9 @@ describe("DataComponent", () => {
                 DynamicTableModule,
                 EditorModule,
                 MatProgressBarModule,
+                CdkAccordionModule,
             ],
-            declarations: [DataComponent, LoadMoreComponent, RequestTimerComponent],
+            declarations: [DataComponent, LoadMoreComponent, RequestTimerComponent, SqlEditorComponent],
         }).compileComponents();
         fixture = TestBed.createComponent(DataComponent);
         datasetSubsService = TestBed.inject(DatasetSubscriptionsService);