+++ src/app/dataset-view/additional-components/data-component/data.component.ts
@@ -9,10 +9,8 @@ import {
 } from "@angular/core";
 import { Location } from "@angular/common";
 import { Observable, map, tap } from "rxjs";
-
 import AppValues from "src/app/common/app.values";
 import DataTabValues from "./mock.data";
-
 import { OffsetInterval } from "../../../api/kamu.graphql.interface";
 import { DataSqlErrorUpdate, DataUpdate, OverviewUpdate } from "src/app/dataset-view/dataset.subscriptions.interface";
 import { DataRow, DatasetRequestBySql } from "../../../interface/dataset.interface";