+++ src/app/dataset-view/additional-components/data-component/request-timer/request-timer.component.spec.ts
@@ -1,4 +1,4 @@
-import { ComponentFixture, TestBed } from "@angular/core/testing";
+import { ComponentFixture, TestBed, fakeAsync, flush, tick } from "@angular/core/testing";
 import { RequestTimerComponent } from "./request-timer.component";
 import { findElementByDataTestId } from "src/app/common/base-test.helpers.spec";
 
@@ -19,21 +19,31 @@ describe("RequestTimerComponent", () => {
         expect(component).toBeTruthy();
     });
 
-    it("should check show duration time when milliseconds < 0", () => {
-        component.resultTime = 0;
-        component.class = "text-bg-success";
+    it("should check rennder result time", fakeAsync(() => {
+        const testClass = "mock-class";
+        const timeElement = findElementByDataTestId(fixture, "duration-request-time");
+        component.class = testClass;
+        const resultTime = "00:00:01.00";
+        component.ngOnChanges({
+            sqlLoading: {
+                previousValue: undefined,
+                currentValue: true,
+                firstChange: true,
+                isFirstChange: () => true,
+            },
+        });
+        tick(1000);
+        component.ngOnChanges({
+            sqlLoading: {
+                previousValue: true,
+                currentValue: false,
+                firstChange: false,
+                isFirstChange: () => false,
+            },
+        });
         fixture.detectChanges();
-        const durationTimeElement = findElementByDataTestId(fixture, "duration-time");
-        expect(durationTimeElement?.textContent?.trim()).toEqual("00:00:00.000");
-        expect(durationTimeElement?.classList).toContain("text-bg-success");
-    });
-
-    it("should check show duration time when milliseconds > 0", () => {
-        component.resultTime = 1000;
-        component.class = "text-bg-success";
-        fixture.detectChanges();
-        const durationTimeElement = findElementByDataTestId(fixture, "duration-time");
-        expect(durationTimeElement?.textContent?.trim()).toEqual("00:00:01.00");
-        expect(durationTimeElement?.classList).toContain("text-bg-success");
-    });
+        expect(timeElement?.textContent?.trim()).toEqual(resultTime);
+        expect(timeElement?.classList).toContain(testClass);
+        flush();
+    }));
 });